+++ subprojects/language-objective-c/src/test/groovy/dev/nokee/language/objectivec/ObjectiveCSourceSetLegacyTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.language.c;
+package dev.nokee.language.objectivec;
 
 import dev.nokee.internal.testing.util.ProjectTestUtils;
 import dev.nokee.language.base.internal.LanguageSourceSetIdentifier;
-import dev.nokee.language.base.testers.LanguageSourceSetTester;
-import dev.nokee.language.c.internal.plugins.CSourceSetRegistrationFactory;
+import dev.nokee.language.base.testers.LanguageSourceSetLegacyTester;
+import dev.nokee.language.objectivec.internal.plugins.ObjectiveCSourceSetRegistrationFactory;
 import dev.nokee.model.internal.registry.ModelRegistry;
 import lombok.val;
 
@@ -26,22 +26,22 @@ import java.io.File;
 
 import static dev.nokee.model.internal.ProjectIdentifier.ofRootProject;
 
-class CSourceSetTest extends LanguageSourceSetTester<CSourceSet> {
+class ObjectiveCSourceSetLegacyTest extends LanguageSourceSetLegacyTester<ObjectiveCSourceSet> {
 	@Override
-	public CSourceSet createSubject() {
+	public ObjectiveCSourceSet createSubject() {
 		val project = ProjectTestUtils.rootProject();
-		project.getPluginManager().apply("dev.nokee.c-language-base");
+		project.getPluginManager().apply("dev.nokee.objective-c-language-base");
 		val registry = project.getExtensions().getByType(ModelRegistry.class);
-		val factory = project.getExtensions().getByType(CSourceSetRegistrationFactory.class);
-		return registry.register(factory.create(LanguageSourceSetIdentifier.of(ofRootProject(), "test"))).as(CSourceSet.class).get();
+		val factory = project.getExtensions().getByType(ObjectiveCSourceSetRegistrationFactory.class);
+		return registry.register(factory.create(LanguageSourceSetIdentifier.of(ofRootProject(), "test"))).as(ObjectiveCSourceSet.class).get();
 	}
 
 	@Override
-	public CSourceSet createSubject(File temporaryDirectory) {
+	public ObjectiveCSourceSet createSubject(File temporaryDirectory) {
 		val project = ProjectTestUtils.createRootProject(temporaryDirectory);
-		project.getPluginManager().apply("dev.nokee.c-language-base");
+		project.getPluginManager().apply("dev.nokee.objective-c-language-base");
 		val registry = project.getExtensions().getByType(ModelRegistry.class);
-		val factory = project.getExtensions().getByType(CSourceSetRegistrationFactory.class);
-		return registry.register(factory.create(LanguageSourceSetIdentifier.of(ofRootProject(), "test"))).as(CSourceSet.class).get();
+		val factory = project.getExtensions().getByType(ObjectiveCSourceSetRegistrationFactory.class);
+		return registry.register(factory.create(LanguageSourceSetIdentifier.of(ofRootProject(), "test"))).as(ObjectiveCSourceSet.class).get();
 	}
 }
